Create `useCurrentDate` hook
Created hook that return current date in unix timestamp updated every seconds
